+++ capa/features/freeze.py
@@ -53,12 +53,12 @@ import zlib
 import logging
 from typing import Dict, Type
 
+import capa.helpers
 import capa.features.file
 import capa.features.insn
 import capa.features.common
 import capa.features.basicblock
 import capa.features.extractors.base_extractor
-from capa.helpers import hex
 from capa.features.common import Feature
 
 logger = logging.getLogger(__name__)
@@ -87,6 +87,7 @@ def dumps(extractor):
     returns:
       str: the serialized features.
     """
+    hex = capa.helpers.hex
     ret = {
         "version": 1,
         "base address": extractor.get_base_address(),